requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.4   #Update from 3.2.11 to 3.2.12 to 3.2.13 to 4.1 to 4.0.4
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
django-environ==0.8.1
graphviz==0.20  # Neu für quality_prediction_sample_tree

# sklearn== oder auf scikit-learn umstellen

environ==1.0
djangorestframework==3.13.1
markdown==3.3.7 # Markdown support for the browsable API
django-filter==22.1 #Filterung support